+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- add support for nested/recursive includes with hierarchical namespaces (ec960ba)
+
 - parse namespaces from includes section instead of filename regex (f6be2fd)
 
 ### Changed